changed double to float + fixed missing re-calculating of best action after re-building tree

+++ src/main/java/org/game/model/AI/AIPlayerBehavior.java
@@ -6,5 +6,5 @@ public interface AIPlayerBehavior {
     void startGame();
     void endGame();
     void startActionExecution();
-    Double calculatePriority(Coordinate goal);
+    float calculatePriority(Coordinate goal);
 }
